+++ asdf/_tests/test_deprecated.py
@@ -146,3 +146,9 @@ def test_AsdfFile_ignore_implicit_conversion_deprecation(value):
 def test_walk_and_modify_ignore_implicit_conversion_deprecation(value):
     with pytest.warns(AsdfDeprecationWarning, match="ignore_implicit_conversion is deprecated"):
         asdf.treeutil.walk_and_modify({}, lambda obj: obj, ignore_implicit_conversion=value)
+
+
+@pytest.mark.parametrize("value", [True, False])
+def test_ignore_version_mismatch_deprecation(value):
+    with pytest.warns(AsdfDeprecationWarning, match="ignore_version_mismatch is deprecated"):
+        asdf.AsdfFile({}, ignore_version_mismatch=value)